Deploy error checking
Added error checking to build and deploy scripts, which stop execution when hit. Goal is to use these scripts to test deployment on a test server (using the production database). If all goes well, then upgrading production should work without any problems

+++ scripts/deploy.sh
@@ -83,10 +83,30 @@ if [ ! "$(docker ps -q -f name=nginx-proxy)" ] || [ ! "$(docker ps -q -f name=ng
 fi
 
 # Copy current database and build to a safe location, under a temporary directory.
-info "Copying current database and build to a safe location, under the same temporary directory that the new build is using..."
 cd "${PROJECT_LOCATION}"
-cp -rp ./data/postgres /tmp/${VERSION}
-cp -rp ./packages/ui/build/* /tmp/${VERSION}/old-build
+DB_TMP="/var/tmp/${VERSION}/postgres"
+BUILD_TMP="/var/tmp/${VERSION}/old-build"
+# Don't copy database if it already exists in /var/tmp
+if [ ! -d "${DB_TMP}" ]; then
+    info "Copying old database to ${DB_TMP}"
+    cp -rp ./data/postgres "${DB_TMP}"
+    if [ $? -ne 0 ]; then
+        error "Could not copy database to ${DB_TMP}"
+        exit 1
+    fi
+else 
+    info "Old database already exists at ${DB_TMP}, so not copying it"
+fi
+if [ ! -d "${BUILD_TMP}" ]; then
+    info "Moving old build to ${BUILD_TMP}"
+    mv -f ./packages/ui/build "${BUILD_TMP}"
+    if [ $? -ne 0 ]; then
+        error "Could not move build to ${BUILD_TMP}"
+        exit 1
+    fi
+else
+    info "Old build already exists at ${BUILD_TMP}, so not moving it"
+fi
 
 # Stop docker containers
 info "Stopping docker containers..."
@@ -102,11 +122,19 @@ info "Cleaning yarn cache..."
 yarn cache clean
 info "Running yarn..."
 yarn
+if [ $? -ne 0 ]; then
+    error "Yarn failed"
+    exit 1
+fi
 
-# Move build created by build.sh to the correct location.
-info "Moving new build to correct location..."
+# Move and decompress build created by build.sh to the correct location.
+info "Moving and decompressing new build to correct location..."
 rm -rf ./packages/ui/build
-cp -rp /tmp/${VERSION}/build ./packages/ui/build
+tar -xzf /var/tmp/${VERSION}/build.tar.gz -C ./packages/ui
+if [ $? -ne 0 ]; then
+    error "Could not move and decompress build to correct location"
+    exit 1
+fi
 
 # Restart docker containers.
 info "Restarting docker containers..."
